+++ CHANGELOG.md
@@ -4,7 +4,17 @@ This project does its best to adhere to [Semantic Versioning](http://semver.org/
 
 
 --------
-### [0.11.1](N/A) - 2016-12-21
+### [0.12.0](N/A) - 2017-01-22
+Major changes to use new TypeScript 2.0 features
+#### Changed
+* hasMatchingProperties() passes a second parameter, the property name, to the 'filter' function parameter
+* Stronger parameter types (mostly `keyof` changes) for:
+  * Arrays: binarySearch(), findMatchingProps(), firstProp(), pluck(), hasItems(), indexOfProp(), lastIndexOfProp(), setAllProp()
+  * Objects: hasAnyNonFalseyProps(), hasAnyNonNullProps(), hasNonNullProps(), hasMatchingProps(), clone(), map(), toArray()
+
+
+--------
+### [0.11.1](https://github.com/TeamworkGuy2/ts-mortar/commit/a5917701fe457a92b06a8b7f23f5e1c2a5f040c3) - 2016-12-21
 #### Changed
 Minor changes for TypeScript 2.0 compatibility